move the only method used in ModelPublication into JournalUtils
